+++ docs/changes.rst
@@ -87,6 +87,8 @@ Released: not yet
 * The zhmcclient package no longer adds a NullHandler to the Python root
   logger (but still to the zhmcclient.api/.hmc loggers).
 
+* Added a function test concept that tests against a real HMC.
+
 **Known issues:**
 
 * See `list of open issues`_.